mlops/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# ML Training Dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0

# XAI Libraries
shap>=0.42.0
lime>=0.2.0

# Data Versioning
dvc>=3.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

apache-airflow-providers-http
apache-airflow-providers-postgres

kaggle
kagglehub[pandas-datasets]
pandas
numpy

